% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_with_criteria}
\alias{aggregate_with_criteria}
\alias{mean_with_criteria}
\alias{min_with_criteria}
\alias{max_with_criteria}
\alias{sum_with_criteria}
\title{Calculate an aggregated value taking into account certain exclusion cirteria}
\usage{
aggregate_with_criteria(x, max_missing = 0, max_consecutive = 0, fun = mean)

mean_with_criteria(x, max_missing = 0, max_consecutive = 0)

min_with_criteria(x, max_missing = 0, max_consecutive = 0)

max_with_criteria(x, max_missing = 0, max_consecutive = 0)

sum_with_criteria(x, max_missing = 0, max_consecutive = 0)
}
\arguments{
\item{x}{Numeric values}

\item{max_missing}{Numeric. Maximum number or proportion of missing values in
x. If any number of missing values is allowed set to 1 (100\%) or NULL.}

\item{max_consecutive}{Numeric. Maximum number or proportion of consecutive
missing values in x. If any number of missing values is allowed set to 1
(100\%) or NULL.}

\item{fun}{Function. Function used to aggregate values usually, mean, min,
max, or sum}
}
\value{
NA if the criteria arent meet the mean of x otherwise
}
\description{
Often one needs to calulate the mean accepting certain number of missing
values. For example, the World Meteorological Organisation specifies that
monthly values must only be calculated if there is up to 10 missing days and
less than 5 are consecutive. This function makes it easy to compute the
monthly value by specifing these thresholds
}
\section{Functions}{
\itemize{
\item \code{mean_with_criteria()}: Mean with criteria

\item \code{min_with_criteria()}: Minimum with criteria

\item \code{max_with_criteria()}: Maximum with criteria

\item \code{sum_with_criteria()}: Sum with criteria

}}
