% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{all_columns_to_snakecase}
\alias{all_columns_to_snakecase}
\title{Make all columns of a data frame snakecase.}
\usage{
all_columns_to_snakecase(x)
}
\arguments{
\item{x}{The input data frame.}
}
\description{
This function modifies the input data frame such that the names of every
column are in snake case. Internally, it uses the
\code{snakecase::to_snake_case}.
}
\examples{
x <- data.frame(notSnakecase = 1:10)
all_columns_to_snakecase(x)
}
