% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{are_all_the_same}
\alias{are_all_the_same}
\title{Are all values on x the same?}
\usage{
are_all_the_same(x)
}
\arguments{
\item{x}{the values}
}
\value{
TRUE if all values of x are the same, FALSE otherwise
}
\description{
Are all values on x the same?
}
