% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{aws_profile_configured}
\alias{aws_profile_configured}
\title{Test whether an AWS profile is configured}
\usage{
aws_profile_configured()
}
\value{
Logical scalar
}
\description{
Test whether an AWS profile is configured
}
\keyword{internal}
