% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly.R
\name{calc_annual_anomaly}
\alias{calc_annual_anomaly}
\title{Calculate annual anomaly}
\usage{
calc_annual_anomaly(x, year, period = NULL)
}
\arguments{
\item{x}{metric. A vector}

\item{year}{the year in which the metric was observed. A vector of the same
length as x}

\item{period}{the reference period of the anomaly. A vector of length 2 or
NULL. If a vector, the first element indicates the beginning of the
reference period and the second the end of the reference period. If NULL
(the default) the reference period goes from the first to the last year}
}
\value{
a vector with the anomaly of x
}
\description{
Calculate annual anomaly
}
\examples{

temperature_frame <- tibble::tibble(temperature = runif(100, 10, 20),
                                    year = 1901:2000)
reference_period <- c(1961, 1990)
temperature_frame \%>\%
dplyr::mutate(anomaly = calc_annual_anomaly(temperature,
                                           year,
                                           reference_period))
}
