% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{check_aws_mfa_access}
\alias{check_aws_mfa_access}
\title{Check AWS authentication}
\usage{
check_aws_mfa_access()
}
\arguments{
\item{bucket_name}{AWS S3 bucket}
}
\value{
TRUE if authentication succeeds
}
\description{
Validates that the configured AWS profile can access the specified bucket.
}
\keyword{internal}
