% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{get_bucket_version_df}
\alias{get_bucket_version_df}
\title{List bucket contents with versions}
\usage{
get_bucket_version_df(bucket_name = mfe_datalake_bucket)
}
\arguments{
\item{bucket_name}{Name of the bucket to connect. By default, it uses the
Ministry for the Environment data lake for environmental reporting
"mfedlkinput".}
}
\value{
A data frame with metadata about all of the versions of objects in a
  bucket.
}
\description{
Returns metadata about all of the versions of objects in a bucket.
}
