% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{get_metadata}
\alias{get_metadata}
\title{Retrieve metadata from an object in the data lake. This function is designed to retrieve the attribute information that is attatched to .RDS file types in the write_to_datalake() function.}
\usage{
get_metadata(..., from_datalake = T, data = NULL)
}
\arguments{
\item{...}{Key words for an object in the lake}
}
\value{
TRUE if it succeeded and FALSE if it failed
}
\description{
Retrieve metadata from an object in the data lake. This function is designed to retrieve the attribute information that is attatched to .RDS file types in the write_to_datalake() function.
}
