% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{get_reference_rainfall}
\alias{get_reference_rainfall}
\title{Calculate reference rainfall for a reference period/climate normal}
\usage{
get_reference_rainfall(rainfall, date = NULL, reference_period = NULL,
  percentile = 95L)
}
\arguments{
\item{rainfall}{vector with rainfall values}

\item{date}{year for each rain value. Must be the same lentgth of rainfall}

\item{reference_period}{vector with the beginning and end of the reference
period}

\item{percentile}{percentile for the reference period. Defaults to the 95th
percentile}
}
\value{
the 95th percentile for the reference period.
}
\description{
The reference rainfall corresponds to the (usually) 95th percentile of
rainfall within the reference period.
}
\examples{
#' library(dplyr)
# Simulate one measurement of rain per day for 10 years
rainfall <- rlnorm(10*365)
years <- rep(2001:2010, each = 365)
rain_data <- tibble(rainfall = rainfall, year = years)
# We chose a climate normal
climate_normal <- c(2001, 2005)

rain_data \%>\%
# calculate reference rainfall
  mutate(ref = get_reference_rainfall(rainfall,
                                             year,
                                             climate_normal,
                                             percentile = 95L)) \%>\%
# calculate prorportion of rainfall above reference
  group_by(year) \%>\%
  summarise(prop_above = rainfall_above_reference(rainfall, ref))
}
\seealso{
Other rainfall functions: \code{\link{rainfall_above_reference}}
}
\concept{rainfall functions}
