% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{get_season}
\alias{get_season}
\title{Get the season of a date-time}
\usage{
get_season(x, with_year = TRUE)
}
\arguments{
\item{x}{a date-time object of class POSIXct, POSIXlt, Date, chron, yearmon, yearqtr,
zoo, zooreg, timeDate, xts, its, ti, jul, timeSeries, fts or anything else that can
be converted with as.POSIXlt}

\item{with_year}{logical indicating whether or not to include the season's year/}
}
\value{
a string with the season or the year and the season
}
\description{
Gets the metereological season of a date or date-time object. Days in
December belong to the summer of the next year. For example December 24, 2010
belongs to summer 2011.
}
\examples{
 x <- lubridate::ymd("2010-01-15") + seq(0, 365*2, length.out = 25)
 get_season(x, with_year = TRUE)
 get_season(x, with_year = FALSE)
}
