% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridify_mb_data.R
\name{gridify_mb_data}
\alias{gridify_mb_data}
\title{Gridify meshblock data}
\usage{
gridify_mb_data(data, shp, grid)
}
\arguments{
\item{data}{A data object with meshblock id variable called id and 1 numeric linecode variable called var. Required input.}

\item{shp}{A sf object of the meshblock shapes with meshblock id variable called id. Use the year after the APS year for a better join. Required input.}

\item{grid}{A sf object of a grid. E.g. er.helpers::nz_grid_hex_346. Required input.}
}
\value{
A sf object with estimated grid totals and densities for the var in the input data.
}
\description{
Converts meshblock data, a meshblock sf object and a grid sf object to grid sf object with density.
}
\examples{
\dontrun{
library(dplyr)
library(sf)
library(simplevis)
library(GMS)

agdir <- "~/Network-Shares/U-Drive-SAS-03BAU/IndLabr/Bus Infra & Performance/Agriculture/Ag_Secure/"

data <- haven::read_sas(paste0(agdir, "2017/all_linecodes_perturbed_final.sas7bdat")) \%>\%
  select(id = CAR_Meshblock, var = APS_Val_IrriTotArea) #data with cols id and var

shp <- get_feature_class("MB2018_V1_00",  epsg = 2193) \%>\%
  select(id = mb2018_v1_00)  #shp with col id

grid_irrigated_land_17 <- gridify_mb_data(
  data = data,
  shp = shp,
  grid = er.helpers::nz_grid_hex_346)

leaflet_sf_col(grid_irrigated_land_17,
               grid_density_per_km2,
               col_method = "bin", col_cuts = c(0, 3, 15, 30, 45, 60, Inf))
}
}
