% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{in_interval}
\alias{in_interval}
\title{Wether a value is within a specified interval}
\usage{
in_interval(x, lower, upper, left_open, right_open)
}
\arguments{
\item{x}{the value to test}

\item{lower}{lower limit}

\item{upper}{upper limit}

\item{left_open}{whether the lower limit is an open interval}

\item{right_open}{whether the upper limit is an open interval}
}
\value{
logical
}
\description{
Wether a value is within a specified interval
}
