% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipcc_likelihood_scale}
\alias{ipcc_likelihood_scale}
\title{IPCC likelihood scale}
\format{
A data frame with 9 rows and 5 variables:
\describe{
  \item{term}{likelihood term}
  \item{left_break}{lower probability limit for the term}
  \item{right_break}{upper probability limit for the term}
  \item{left_open}{whether the lower limit is an open interval}
  \item{right_open}{whether the upper limit is an open interval}
}
}
\source{
Mastrandrea, M.D., C.B. Field, T.F. Stocker, O. Edenhofer, K.L. Ebi,
  D.J. Frame, H. Held, E. Kriegler, K.J. Mach, P.R. Matschoss, G.-K.
  Plattner, G.W. Yohe, and F.W. Zwiers, 2010: Guidance Note for Lead Authors
  of the IPCC Fifth Assessment Report on Consistent Treatment of
  Uncertainties. Intergovernmental Panel on Climate Change (IPCC). Available
  at \url{http://www.ipcc.ch}
}
\usage{
ipcc_likelihood_scale
}
\description{
A dataset containing the Intergovernmental Panel on Climate Change (IPCC)
likelihood scale
}
\keyword{datasets}
