% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_shiny_in_background.R
\name{launch_shiny_in_background}
\alias{launch_shiny_in_background}
\title{Launch a shiny app in the background}
\usage{
launch_shiny_in_background(path = ".", port = NULL)
}
\arguments{
\item{path}{Path of the shiny app, defaults to working directory}

\item{port}{Port to be used for the connection. If NULL uses a randomly
generated port between 1000 and 9999}
}
\value{
The URL of the running app
}
\description{
Launch a shiny app in the background
}
\examples{

launch_shiny_in_backrgound("app/", port = 6589)

}
