% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{linear_model}
\alias{linear_model}
\title{Linear model for trend analysis}
\usage{
linear_model(x, y = NULL, conf_level = 0.95)
}
\arguments{
\item{x}{numeric vector}

\item{y}{optional time variable, converted to numeric. If its not provided it
will be assumed that all values in X are sequential, regularly measured,
and there are no gaps in the measurements.}
}
\value{
a tidy data frame with the model results
}
\description{
Performs a linear model for
}
\examples{
x <- runif(100) * 1:100
linear_model(x)
# If measurements were for example taken daily
linear_model(x, y = Sys.Date() + (1:100))
}
