% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{mann_kendall}
\alias{mann_kendall}
\title{Mann-Kendall Trend Test}
\usage{
mann_kendall(
  x,
  alternative = c("greater", "two.sided", "less"),
  continuity = TRUE
)
}
\arguments{
\item{x}{a vector of class "numeric" or a time series object of class "ts"}

\item{alternative}{the alternative hypothesis, defaults to \code{two.sided}}

\item{continuity}{logical, indicates whether a continuity correction
should be applied, defaults to \code{TRUE}.}
}
\value{
a tidy data frame with the test results
}
\description{
Performs the Mann-Kendall trend test. This function is a wrapper around
\code{\link[trend]{mk.test}}. See documentation there for more details about
the calculation
}
\examples{
x <- runif(100) * 1:100
mann_kendall(x, alternative = "two.sided")
}
