% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{metadata_to_table}
\alias{metadata_to_table}
\title{Metadata to table}
\usage{
metadata_to_table(df, remove_names = c("row.names"))
}
\arguments{
\item{df}{- df object with attributes}

\item{remove_names}{- Remove names e.g. row.names or groups}
}
\value{
tibble with name and values of attributes extracted from df
}
\description{
Metadata to table
}
