% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\docType{data}
\name{mfe_datalake_bucket}
\alias{mfe_datalake_bucket}
\title{MfE Default bucket}
\format{
An object of class \code{character} of length 1.
}
\usage{
mfe_datalake_bucket
}
\description{
MfE Default bucket
}
\keyword{datasets}
