% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz}
\alias{nz}
\title{New Zealand coastline.}
\format{
An \code{sf} object.
}
\usage{
nz
}
\description{
Simplified New Zealand coastline boundary, excluding the Chatham Islands.
}
\keyword{datasets}
