% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz_grid_hex_346}
\alias{nz_grid_hex_346}
\title{Hexagonal grid sf object with hexagons of 346km2.}
\format{
An \code{sf} object.

An \code{sf} object.
}
\usage{
nz_grid_hex_346

nz_grid_hex_346
}
\description{
Hexagonal grid sf object with hexagons of 346km2.

Hexagonal grid sf object with hexagons of 346km2.
}
\keyword{datasets}
