% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{order_likelihood_levels}
\alias{order_likelihood_levels}
\title{Orders the levels of a likelihood category factor}
\usage{
order_likelihood_levels(x)
}
\arguments{
\item{x}{the unordered factor. Levels must be one of those found in
er.helpers:::likelihood_terms}
}
\value{
a factor with ordered levels
}
\description{
Orders the levels of a likelihood category factor
}
\examples{
\dontrun{
library(ggplot2)
library(tibble)
trends <- tribble(
  ~ trend, ~n,
  "Indeterminate", 2,
  "Very likely worsening", 3,
  "Likely improving", 2)
# unordered plot
qplot(trend, n, fill = trend, data = trends, geom = "col")

ordered_trends <- dplyr::mutate(trends,
                                trend = order_likelihood_levels(trend))
# nice ordered plot
qplot(trend, n, fill = trend,  data = ordered_trends, geom = "col")
}
}
\seealso{
Other likelihood functions:
\code{\link[=get_likelihood_category]{get_likelihood_category()}},
\code{\link[=simplify_likelihood_levels]{simplify_likelihood_levels()}}
}
\concept{likelihood functions}
