% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{order_season_levels}
\alias{order_season_levels}
\title{Order season levels}
\usage{
order_season_levels(x)
}
\arguments{
\item{x}{character vector. If a factor will be coerced to character first}
}
\value{
a factor with ordered levels for season
}
\description{
Standardise season as a factor with ordered seasons. If an "annual" category
is present it includes at the end of the other seasons. Season names are
converted to sentence case
}
\examples{
library(ggplot2)
library(tibble)
seasons <- tribble(
  ~ season, ~n,
  "autum", 1,
  "summer", 2,
  "winter", 3,
  "spring", 1,
  "annual", 2)
# unordered ugly plot
qplot(season, n, fill = season, data = seasons, geom = "col")

ordered_seasons <- dplyr::mutate(seasons,
                                season = order_season_levels(season))
# nice ordered plot
qplot(season, n, fill = season, data = ordered_seasons, geom = "col")

}
