% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{pal_ea19}
\alias{pal_ea19}
\title{Colour palette for a graph with a nominal categorical variable.}
\usage{
pal_ea19
}
\value{
A vector of hex codes.
}
\description{
Colour palette for a graph with a nominal categorical variable.
}
\examples{
scales::show_col(pal_ea19)
}
