% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{pal_ea19_nztcs_c}
\alias{pal_ea19_nztcs_c}
\title{Colour palette for a NZTCS category graph.}
\usage{
pal_ea19_nztcs_c
}
\value{
A vector of hex codes.
}
\description{
A colour palette used for depicting categories in the NZ conservation threat status.
}
\examples{
scales::show_col(pal_ea19_nztcs_c)
}
