% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{pal_ea19_trend5}
\alias{pal_ea19_trend5}
\title{Colour palette for a graph of a ordinal categorical trend variable with 5 values.}
\usage{
pal_ea19_trend5
}
\value{
A vector of hex codes.
}
\description{
Colour palette for a graph of a ordinal categorical trend variable with 5 values.
}
\examples{
scales::show_col(pal_ea19_trend5)
}
