% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{pal_point_set1}
\alias{pal_point_set1}
\title{Colour palette for categorical variables for points on a map etc.}
\usage{
pal_point_set1
}
\value{
A vector of hex codes.
}
\description{
Colour palette for categorical variables.
}
\examples{
scales::show_col(pal_point_set1)
}
