% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{pal_snz}
\alias{pal_snz}
\title{Colour palette for a graph with a nominal categorical variable.}
\usage{
pal_snz
}
\value{
A vector of hex codes.
}
\description{
Colour palette for a graph with a nominal categorical variable.
}
\examples{
scales::show_col(pal_snz)
}
