% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{pal_snz_alpha2}
\alias{pal_snz_alpha2}
\title{Colour palette for a graph that compares a current year to 1 past year.}
\usage{
pal_snz_alpha2
}
\value{
A vector of hex codes. Uses the first colour and a 30% tint.
}
\description{
Colour palette for a graph that compares a current year to 1 past year.
}
\examples{
scales::show_col(pal_snz_alpha2)
}
