% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{pal_snz_continuous}
\alias{pal_snz_continuous}
\title{Colour palette for a graph with a continuous variable.}
\usage{
pal_snz_continuous
}
\value{
A vector of hex codes.
}
\description{
Colour palette for a graph with a continuous variable.
}
\examples{
scales::show_col(pal_snz)
}
