% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{pal_snz_nztcs_c}
\alias{pal_snz_nztcs_c}
\title{Colour palette for a NZTCS category graph.}
\usage{
pal_snz_nztcs_c
}
\value{
A vector of hex codes.
}
\description{
Colour palette for a NZTCS category graph.
}
\examples{
scales::show_col(pal_snz_nztcs_c)
}
