% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{pal_snz_nztcs_s}
\alias{pal_snz_nztcs_s}
\title{Colour palette for a NZTCS subcategory graph.}
\usage{
pal_snz_nztcs_s
}
\value{
A vector of hex codes.
}
\description{
A colour palette used for depicting subcategories in the NZ conservation threat status.
}
\examples{
scales::show_col(pal_snz_nztcs_s)
}
