% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{pal_snz_trend3}
\alias{pal_snz_trend3}
\title{Colour palette for a graph of a ordinal categorical trend variable with 3 values.}
\usage{
pal_snz_trend3
}
\value{
A vector of hex codes.
}
\description{
Colour palette for a graph of a ordinal categorical trend variable with 3 values.
}
\examples{
scales::show_col(pal_snz_trend3)
}
