% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{prepare_pattern}
\alias{prepare_pattern}
\title{Prepare a pattern for look in the data lake and use with str_detect}
\usage{
prepare_pattern(pattern)
}
\arguments{
\item{pattern}{- list of patterns from search datalake}
}
\value{
list of patterns using standard unicode collation rules
}
\description{
Prepare a pattern for look in the data lake and use with str_detect
}
