% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{rainfall_above_reference}
\alias{rainfall_above_reference}
\title{Get proportion of rainfall above reference}
\usage{
rainfall_above_reference(rainfall, reference_rainfall)
}
\arguments{
\item{rainfall}{vector with rainfall values}

\item{reference_rainfall}{reference value of rainfall}
}
\value{
a value between 0 and 1
}
\description{
Calculates the proportion of rainfall above the refrerence_rainfall value.
Usually the reference corresponds to the 95th percentile of a climate
normal/reference period. The proportion is often calculated in an annual
basis and is a measure of the proportion of annual total rain that falls in
intense events. This measure provides information about the importance of
intense rainfall events for total annual rainfall.
}
\examples{
library(dplyr)
# Simulate one measurement of rain per day for 10 years
rainfall <- rlnorm(10*365)
years <- rep(2001:2010, each = 365)
rain_data <- tibble(rainfall = rainfall, year = years)
# We chose a climate normal
climate_normal <- c(2001, 2005)

rain_data \%>\%
# calculate reference rainfall
  mutate(ref = get_reference_rainfall(rainfall,
                                             year,
                                             climate_normal,
                                             percentile = 95L)) \%>\%
# calculate prorportion of rainfall above reference
  group_by(year) \%>\%
  summarise(prop_above = rainfall_above_reference(rainfall, ref))
}
\seealso{
Other rainfall functions: \code{\link{get_reference_rainfall}}
}
\concept{rainfall functions}
