% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{read_aws_csv}
\alias{read_aws_csv}
\title{Read a CSV file stored in an AWS S3 bucket with enabled MFA authentication.}
\usage{
read_aws_csv(s3_path, bucket_name = mfe_datalake_bucket, version = NULL, ...)
}
\arguments{
\item{s3_path}{The filename of the desired CSV in the S3 bucket including the
full path}

\item{bucket_name}{S3 bucket name}

\item{version}{VersionId of the object key desired. Can be retrieved using
\code{\link{get_bucket_version_df}}}

\item{...}{Other arguments passed to the reading_function}
}
\value{
A `\code{\link[tibble]{tibble}}
}
\description{
This function get the specified object from an AWS S3 bucket and reads it
using \code{\link[readr]{read_csv}}. It keeps the CSV in memory and,
therefore, it avoids the unintended consequences of saving the file in the
disk.
}
\examples{
\dontrun{
setup_aws_access()
csv_object_path <- "freshwater/2020/raw/urban_stream_water_quality_state.csv"
read_aws_csv(csv_object_path)
}
}
