% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{read_excel_datalake}
\alias{read_excel_datalake}
\title{Read a excel file stored in an AWS S3 bucket.}
\usage{
read_excel_datalake(s3_path, bucket_name = mfe_datalake_bucket,
  version = NULL, ...)
}
\arguments{
\item{s3_path}{The filename of the desired excel in the S3 bucket including the
full path}

\item{bucket_name}{Name of the bucket to connect. By default, it uses the
Ministry for the Environment data lake for environmental reporting
"mfedlkinput".}

\item{version}{VersionId of the object key desired. Can be retrieved using
\code{\link{get_bucket_version_df}}}

\item{...}{Other arguments passed to the reading_function}
}
\value{
A `\code{\link[tibble]{tibble}}
}
\description{
This function get the specified object from an AWS S3 bucket and reads it
using \code{\link[readxl]{read_excel}}. It keeps the excel in memory and,
therefore, it avoids the unintended consequences of saving the file in the
disk.
}
\examples{
\dontrun{
setup_datalake_access()
excel_object_path <- "freshwater/2020/raw/urban_stream_water_quality_state.excel"
read_excel_datalake(excel_object_path)
}
}
