% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{read_from_datalake}
\alias{read_from_datalake}
\title{Read any type file stored in an AWS S3 bucket.}
\usage{
read_from_datalake(..., date = NULL, all_sheets = T)
}
\arguments{
\item{...}{Key terms to search for in the AWS S3 bucket}

\item{all_sheets}{If more than one sheet is present, T = read all sheets into a list, F = default to sheet specified}
}
\description{
This function get the specified object from an AWS S3 bucket and reads it
using functions designed for each relevant file type. If the file is either a .RDS, .csv, or .xlsx; it first downloads the file to a temp directory and,
therefore, it avoids the unintended consequences of saving the file in the
disk. If it is not a recognisable file type, it saves it to the working directory.
This function also uses key search terms and will throw an error if there is more than one file with the search terms used.
}
\examples{
\dontrun{
setup_datalake_access()
read_from_datalake("landcover", "concordance", "lcdb4")
}
}
