% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries.R
\name{rearrange_geometries}
\alias{rearrange_geometries}
\title{rearrange_geometries}
\usage{
rearrange_geometries(string)
}
\arguments{
\item{string}{This is the comma-separated string of lats and corresponding longs (character string expected)}
}
\value{
returns a comma-separated string of lat/long pairs
}
\description{
Takes in geometries arranged with with all the lattitudes first and then all the longitudes
and returns a series of lattitude and longitude pairs.
}
\examples{
geometry <- "c(1,2,3,4,5,6)"
rearrange_geometries(geometry)
}
