% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{round}
\alias{round}
\title{round}
\usage{
round(x, digits)
}
\arguments{
\item{x}{metric. A vector}

\item{digits}{an integer indicating the number of decimal places to be used.}
}
\value{
a vector rounded to the digits specified.
}
\description{
A function used to round numbers using the IEC 60559 standard. base::round doesn't always behave in this manner
(see help round for more information). This function is designed to mask base::round.
}
