% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Round2}
\usage{
round2(x, n)
}
\arguments{
\item{x}{vector}

\item{n}{digits to round to}
}
\description{
Round using conventional rounding, where .5 always gets rounded up.
}
\examples{
round2(0.05, 1)
}
