% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{search_data_lake}
\alias{search_data_lake}
\title{Search for keys in the data lake}
\usage{
search_data_lake(..., bucket_name = mfe_datalake_bucket,
  object_versions = FALSE)
}
\arguments{
\item{bucket_name}{Name of the bucket to connect. By default, it uses the
Ministry for the Environment data lake for environmental reporting
"mfedlkinput".}

\item{object_versions}{}
}
\value{
a data frame with metadata for selected objects
}
\description{
Returns metadata about objects in a bucket. This function is a wrapper to
`aws.s3::get_bucket_df` but filters out the desired Keys. If there are more
than 1000 objects, it makes iterative calls to the AWS S3 API to retrieve the
metadata for all versions. If the function is called from an interactive
session, it invokes a data viewer (\code{\link[utils]{View}}) with the search
results.
}
\examples{
 \dontrun{
search_data_lake("temperature")
# search using regex
search_data_lake(stringr::regex("^a"))
# search tidy datasets for atmosphere and climate 2020
search_data_lake("tidy", "climate", "2020")
}
}
