% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{setup_datalake_access}
\alias{setup_datalake_access}
\title{Set up access to the data lake programatically.}
\usage{
setup_datalake_access(
  cred_csv = "~/credentials.csv",
  bucket_name = mfe_datalake_bucket
)
}
\arguments{
\item{cred_csv}{A csv file with the credentials information. At a minimum it
should have two columns: "Access key ID" and "Secret access key". By
default \code{\link{setup_datalake_access}} looks for this file in the user
directory "~/credentials.csv".}

\item{bucket_name}{Name of the bucket to connect. By default, it uses the
Ministry for the Environment data lake for environmental reporting
"mfedlkinput".}
}
\description{
This function set ups the credentials required to access an AWS S3 bucket
programatically. By default it connects to the Environmental Reporting Data
Lake.
}
\examples{
\dontrun{
setup_datalake_access(cred_csv = "~/credentials.csv", bucket_name = "mfedlkinput")
}
}
