% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{setup_sso_access}
\alias{setup_sso_access}
\title{Set up access to the data lake using an AWS CLI v2 profile}
\usage{
setup_sso_access(profile_name, region = "ap-southeast-2")
}
\arguments{
\item{profile_name}{AWS profile name from ~/.aws/config}

\item{region}{AWS region}

\item{bucket_name}{S3 bucket name}
}
\description{
This function configures the current R session to use an AWS profile that
has been configured via AWS CLI v2.
}
\details{
Users must authenticate beforehand using:

aws sso login --profile <profile_name>
}
\examples{
\dontrun{
aws sso login --profile my_profile

setup_sso_access(
  profile_name = "my_profile"
)
}
}
