% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R, R/scale_trans.R
\name{signed_sqrt_trans}
\alias{signed_sqrt_trans}
\title{Signed square root ggplot scale transformation.}
\usage{
signed_sqrt_trans()

signed_sqrt_trans()
}
\value{
A ggplot scale transformation function.

A ggplot scale transformation.
}
\description{
A signed square root ggplot scale transformation.

A signed square root ggplot scale transformation. This is useful for where there is negative values.
}
