% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{stats_theme_map}
\alias{stats_theme_map}
\title{Stats NZ map theme}
\usage{
stats_theme_map()
}
\value{
A ggplot2 theme object.
}
\description{
A minimal ggplot2 theme for maps styled to Stats NZ conventions.
}
