% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statsnz_likelihood_scale}
\alias{statsnz_likelihood_scale}
\title{Stats NZ likelihood scale}
\format{A data frame with 9 rows and 5 variables:
\describe{
  \item{term}{likelihood term}
  \item{left_break}{lower probability limit for the term}
  \item{right_break}upper probability limit for the term
  \item{left_open}{wether the lower limit is an open interval}
  \item{right_open}{wether the upper limit is an open interval}
}}
\usage{
statsnz_likelihood_scale
}
\description{
A dataset containing the Statistics New Zealand (Stats NZ) likelihood scale.
This scale is a simplification of the IPCC likelihood scale
}
\keyword{datasets}
