% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{table_to_metadata}
\alias{table_to_metadata}
\title{Table to metadata
Adds attributes to df object and returns and prints list of attributes}
\usage{
table_to_metadata(df, metadata)
}
\arguments{
\item{df}{dataframe object}

\item{metadata}{attribute names and values to be added to df object}

\item{retain_row_names}{Logical. defaults to FALSE (row.names attribute will be removed), set to TRUE if row.names attribute to be retained}
}
\value{
list of attribute names and values
}
\description{
Table to metadata
Adds attributes to df object and returns and prints list of attributes
}
