% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{version_list_as_df}
\alias{version_list_as_df}
\title{Coverts list with object version metadata to a data frame}
\usage{
version_list_as_df(versions)
}
\arguments{
\item{versions}{list item as returned by `get_versions_list`}
}
\value{
A data frame
}
\description{
Coverts list with object version metadata to a data frame
}
