% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{write_csv_datalake}
\alias{write_csv_datalake}
\title{Write a CSV file stored in an AWS S3 bucket.}
\usage{
write_csv_datalake(x, s3_path, bucket_name = mfe_datalake_bucket, ...)
}
\arguments{
\item{x}{A data frame to write to the bucket}

\item{s3_path}{The filename of the desired CSV in the S3 bucket including
the full path}

\item{bucket_name}{Name of the bucket to connect. By default, it uses the
Ministry for the Environment data lake for environmental reporting
"mfedlkinput".}

\item{...}{Other arguments passed to the reading_function}
}
\value{

}
\description{
Write a CSV file stored in an AWS S3 bucket.
}
\examples{
setup_datalake_access()
csv_object_path <- "freshwater/2020/raw/urban_stream_water_quality_state.csv"
write_csv_datalake(iris, csv_object_path)
}
