% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{write_excel_datalake}
\alias{write_excel_datalake}
\title{Write a excel file as an object in an AWS S3 bucket.}
\usage{
write_excel_datalake(x, s3_path, bucket_name = mfe_datalake_bucket, ...)
}
\arguments{
\item{x}{A data frame to write to the bucket}

\item{s3_path}{The filename of the desired excel in the S3 bucket including the
full path}

\item{bucket_name}{Name of the bucket to connect. By default, it uses the
Ministry for the Environment data lake for environmental reporting
"mfedlkinput".}

\item{...}{Other arguments passed to the reading_function}
}
\value{
TRUE if it succeeded and FALSE if it failed
}
\description{
Write a excel file as an object in an AWS S3 bucket.
}
\examples{
\dontrun{
setup_datalake_access()
excel_object_path <- "freshwater/2020/raw/urban_stream_water_quality_state.excel"
write_excel_datalake(iris, excel_object_path)
}
}
