% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_unused_levels.R
\name{add_unused_levels}
\alias{add_unused_levels}
\title{Add a row for each unused factor level to ensure plotly displays all levels in the legend.}
\usage{
add_unused_levels(data, var)
}
\arguments{
\item{data}{A tibble, dataframe or sf object. Required input.}

\item{var}{A variable of class factor.}
}
\value{
A tibble, dataframe or sf object. Required input.
}
\description{
Add a row for each unused factor level to ensure plotly displays all levels in the legend.
}
