% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{get_likelihood_category}
\alias{get_likelihood_category}
\title{Title}
\usage{
get_likelihood_category(x, scale = c("statsnz", "ipcc"),
  term_type = c("worsening-improving", "improving-worsening",
  "increasing-decreasing", "likely-unlikely"), x_is = c("probability",
  "percentage"))
}
\arguments{
\item{x_is}{}
}
\description{
Title
}
