% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{metadata_to_table}
\alias{metadata_to_table}
\title{Metadata to table}
\usage{
metadata_to_table(df, retain_row_names = F)
}
\arguments{
\item{df}{- df object with attributes}

\item{retain_row_names}{Logical. defaults to FALSE (row.names attribute will be removed), set to TRUE if row.names attribute to be retained}
}
\value{
tibble with name and values of attributes extracted from df
}
\description{
Metadata to table
}
