% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\docType{data}
\name{pal_continuous_trend}
\alias{pal_continuous_trend}
\title{Colour palette for a map of continuous diverging pal}
\format{
An object of class \code{character} of length 9.
}
\usage{
pal_continuous_trend
}
\value{
A vector of hex codes.
}
\description{
Colour palette for a map of continuous diverging pal for rate of change. Best used in scale_x_gradient...
}
\examples{
scales::show_col(pal_point_trend5)
}
\keyword{datasets}
