% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{read_excel_datalake}
\alias{read_excel_datalake}
\title{Read a excel file stored in an AWS S3 bucket.}
\usage{
read_excel_datalake(
  s3_path,
  bucket_name = mfe_datalake_bucket,
  version = NULL,
  all_sheets = T,
  sheet = 1,
  ...
)
}
\arguments{
\item{s3_path}{The filename of the desired excel in the S3 bucket including the
full path}

\item{bucket_name}{Name of the bucket to connect. By default, it uses the
Ministry for the Environment data lake for environmental reporting
"mfedlkinput".}

\item{version}{VersionId of the object key desired. Can be retrieved using
\code{\link{get_bucket_version_df}}}

\item{all_sheets}{If more than one sheet is present, T = read all sheets into a list, F = default to sheet specified}

\item{sheet}{The sheet number to extract. Defaults to 1.}

\item{...}{Other arguments passed to the reading_function}
}
\value{
A `\code{\link[tibble]{tibble}}
}
\description{
This function get the specified object from an AWS S3 bucket and reads it
using \code{\link[readxl]{read_excel}}. It keeps the excel in memory and,
therefore, it avoids the unintended consequences of saving the file in the
disk.
}
\examples{
\dontrun{
setup_datalake_access()
files <- search_datalake(".x", "land", "2021")$Key
read_excel_datalake(files[2])
read_excel_datalake(files[1], sheet = 2)
}
}
