% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{sen_slope}
\alias{sen_slope}
\title{Sen's slope}
\usage{
sen_slope(x, conf_level = 0.95)
}
\arguments{
\item{x}{numeric vector or a time series object of class "ts"}

\item{conf_level}{confidence level of the test}
}
\value{
a tidy data frame with the test results
}
\description{
#' Performs the Sen's slope trend test. This function is a wrapper around
\code{\link[trend]{sens.slope}}. See documentation there for more details about
the calculation
}
\examples{
x <- runif(100) * 1:100
sen_slope(x)
}
