% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{simplify_likelihood_levels}
\alias{simplify_likelihood_levels}
\title{Simplify likelihood levels}
\usage{
simplify_likelihood_levels(x)
}
\arguments{
\item{x}{The vector with the likelihoods. Must be a factor or a character
string}
}
\value{
A factor with simple categories
}
\description{
Collapse levels on a likelihood factor into smaller categories. For example
"Likely improving" and "Very likely improving" are collapsed into a single
level "Likely or very likely improving".
}
\examples{
p <- seq(0, 1, length.out = 11)

get_likelihood_category(p) \%>\%
  order_likelihood_levels() \%>\%
  simplify_likelihood_levels()

}
\seealso{
Other likelihood functions: 
\code{\link{get_likelihood_category}()},
\code{\link{order_likelihood_levels}()}
}
\concept{likelihood functions}
